# String interpolation optimization miniphase