--- a/tests/run/interpolation-opt.check
+++ b/tests/run/interpolation-opt.check
@@ -3,3 +3,6 @@
 is 3.0
 a1two3.0b
 a1two3.0b
+
+
+Hello World