--- a/tests/run/interpolation-opt.check
+++ b/tests/run/interpolation-opt.check
@@ -1,0 +1,5 @@
+1 plus two\nis 3.0
+1 plus two
+is 3.0
+a1two3.0b
+a1two3.0b