--- a/tests/run/interpolation-opt.check
+++ b/tests/run/interpolation-opt.check
@@ -6,3 +6,5 @@ a1two3.0b
 
 
 Hello World
+Side effect!
+Hello World