--- a/tests/run/interpolation-opt.check
+++ b/tests/run/interpolation-opt.check
@@ -7,4 +7,6 @@ a1two3.0b
 
 Hello World
 Side effect!
-Hello World
+Foo Bar
+Side effect n2!
+Titi Toto