@odersky 2018-03-19
What is the status of the PR? Given the performance improvements it would be great to get this in.