@allanrenucci 2018-04-20
test performance please